redis: introduce ConnWithContext
To enable context propagation for distributed tracing
and monitoring but also for backwards compatibility
with older programs, this change introduces `ConnWithContext`
which is an interface that implements `Conn` but introduces
`Conn` methods that are suffixed with `Context` and take in
`context.Context` as the first argument. For example:
* DoContext(context.Context, ...)
* CloseContext(context.Context)
* SendContext(context.Context, ...)
* FlushContext(context.Context)
* ReceiveContext(context.Context)

If the `Context` suffixed methods are not used, `context.Background()`
is passed by the underlying implementations of `*activeConn` and `*conn`.

Once you've retrieved a `Conn`, from the pool, type assert on it
with `ConnContext` e.g.

    conn := redisPool.GetWithContext(ctx).(redis.ConnWithContext)
    defer conn.CloseContext(ctx)